requirements.txt
argon2-cffi==23.1.0
Flask==3.0.3
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
Flask-PyMongo==2.3.0
Flask-SocketIO==5.3.6
python-dotenv==1.0.1
PyJWT==2.8.0
pymongo==4.8.0
eventlet==0.36.1
gunicorn==22.0.0